apps/ai_service/requirements.txt
django==4.2
graphene-django==3.1.5
django-cors-headers